% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellQuality.R
\name{plotCellQuality}
\alias{plotCellQuality}
\title{plotCellQuality}
\usage{
plotCellQuality(ArchRProj, sample, cutoff_frags = NULL, cutoff_tss = NULL)
}
\arguments{
\item{ArchRProj}{ArchRProject.}

\item{sample}{Sample name.}

\item{cutoff_frags}{Optional additional cutoff for TSS Enrichment to visualize.}
}
\value{
GGplot object of scatter plot (TSS Enrichment over log10 Unique Fragemnts).
}
\description{
This function returns a ggplot object with a scatter plot of TSS Enrichment over log10 Unique Fragments per cell. Default filtering criteria of TSS Enrichment of 4 and minimal number of unique fragments of 1000 are visualized.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{cell}
\keyword{qc}
\keyword{quality}
\keyword{rwire}
