% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoAccessibilityMap.R
\name{plotCoAccessibilityMap}
\alias{plotCoAccessibilityMap}
\title{plotCoAccessibilityMap}
\usage{
plotCoAccessibilityMap(
  ArchRProj = NULL,
  reducedDims = names(ArchRProj@reducedDims)[1],
  region = NULL,
  resolution = 1000,
  ArchRGenome = getArchRGenome()
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{region}{GRanges object of length 1 with region of interest for which co-accessibility should be calculated.}

\item{resolution}{Tile size for which co-accessibility should be calculated. Defaults to 1 kb.}

\item{ArchRGenome}{Used genome and gene annotation.}
}
\value{
Co-accessibility map.
}
\description{
This function returns co-accessibility map of specified genomic region.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{co-accessibility}
\keyword{coAccessibilityMap}
\keyword{map}
\keyword{rwire}
